+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+#### 5th April 2018 - Version 1.0.7
+
+* Tweak: Exclude Content Blocks from search by default
+
+#### 27th March 2018 - Version 1.0.6
+
+* Tweak: Added more info to admin notice if dependent plugins or themes don't fit
+
 #### 4th March 2018 - Version 1.0.5
 
 * Fix: Specify post ID as get_post_type argument to be sure the post is a content block before displaying